refactor(api-v3): split obsolete members into seperate partial class
