+++ source/scripting_v3/GTA.UI/TextElement.cs
@@ -14,7 +14,7 @@ using GTA.Native;
 
 namespace GTA.UI
 {
-    public class TextElement : IWorldDrawableElement
+    public partial class TextElement : IWorldDrawableElement
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="TextElement"/> class used for drawing text on the screen.
@@ -236,20 +236,6 @@ namespace GTA.UI
         public float WrapWidth
         {
             get; set;
-        }
-        /// <summary>
-        /// Gets or sets a value indicating whether the alignment of this <see cref="TextElement" /> is centered.
-        /// See <see cref="Alignment"/>
-        /// </summary>
-        /// <value>
-        ///   <see langword="true" /> if centered; otherwise, <see langword="false" />.
-        /// </value>
-        [Obsolete("`TextElement.Centered` is obsolete because it is redundant and setting the property to false is " +
-            "confusing. Use `TextElement.Alignment` instead."), EditorBrowsable(EditorBrowsableState.Never)]
-        public bool Centered
-        {
-            get => Alignment == Alignment.Center;
-            set => Alignment = value ? Alignment.Center : Alignment.Left;
         }
 
         /// <summary>
