+++ source/scripting_v3/GTA/Entities/Peds/PedComponent.cs
@@ -9,7 +9,7 @@ using GTA.Native;
 
 namespace GTA
 {
-    public sealed class PedComponent : IPedVariation
+    public sealed partial class PedComponent : IPedVariation
     {
         #region Fields
         readonly Ped _ped;
@@ -95,19 +95,6 @@ namespace GTA
         }
 
         public bool HasVariations => Count > 1;
-
-        /// <summary>
-        /// Returns <see langword="true"/> if there are textures for current drawable id (<see cref="Index"/>).
-        /// </summary>
-        [Obsolete("PedComponent.HasTextureVariations is obsolete because it does not make sense " +
-                  "as texture count cannot be determined without specifying both component id and drawable id."),
-        EditorBrowsable(EditorBrowsableState.Never)]
-        public bool HasTextureVariations => Count > 0 && TextureCount > 1;
-
-        [Obsolete("PedComponent.HasAnyVariation is obsolete because it does not make sense " +
-                  "as texture count cannot be determined without specifying both component id and drawable id."),
-        EditorBrowsable(EditorBrowsableState.Never)]
-        public bool HasAnyVariations => HasVariations || HasTextureVariations;
 
         public override string ToString()
         {
