+++ source/scripting_v3/GTA/GameVersionNotSupportedException.Obsolete.cs
@@ -1,0 +1,14 @@
+using System;
+using System.ComponentModel;
+
+namespace GTA
+{
+    public sealed partial class GameVersionNotSupportedException : Exception
+    {
+        [Obsolete("`GameVersionNotSupportedException.MinimumSupportedGameVersion` is deprecated " +
+    "because Script Hook V is deprecating `getGameVersion`, which the property is based on. " +
+    "Use `GameVersionNotSupportedException.MinimumSupportedGameFileVersion` instead.")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public GameVersion MinimumSupportedGameVersion { get; }
+    }
+}
