+++ source/scripting_v3/GTA/GameVersionNotSupportedException.cs
@@ -39,13 +39,8 @@ namespace GTA
     /// </list>
     /// </remarks>
     [Serializable]
-    public sealed class GameVersionNotSupportedException : Exception
+    public sealed partial class GameVersionNotSupportedException : Exception
     {
-        [Obsolete("`GameVersionNotSupportedException.MinimumSupportedGameVersion` is deprecated " +
-            "because Script Hook V is deprecating `getGameVersion`, which the property is based on. " +
-            "Use `GameVersionNotSupportedException.MinimumSupportedGameFileVersion` instead.")]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public GameVersion MinimumSupportedGameVersion { get; }
         public Version MinimumSupportedGameFileVersion { get; }
 
         private Dictionary<Version, GameVersion> _supportedGameVersionEnumMaps
