+++ source/scripting_v3/GTA/Weapons/WeaponComponentCollection.cs
@@ -10,7 +10,7 @@ using System.Linq;
 
 namespace GTA
 {
-    public sealed class WeaponComponentCollection
+    public sealed partial class WeaponComponentCollection
     {
         #region Fields
         readonly Ped _owner;
@@ -344,29 +344,6 @@ namespace GTA
                 }
 
                 if (component.AttachmentPoint is WeaponAttachmentPoint.FlashLaser or WeaponAttachmentPoint.FlashLaser2)
-                {
-                    return component;
-                }
-            }
-            return _invalidComponent;
-        }
-
-        /// <summary>
-        /// Gets the first component of all the components for <see cref="WeaponAttachmentPoint.GunRoot"/>.
-        /// Despite the method name, return value is not guaranteed to a <see cref="WeaponComponent"/> instance that represents the luxury finish component.
-        /// </summary>
-        /// <returns>
-        /// The <see cref="WeaponComponent"/> instance if the first component of all the components for <see cref="WeaponAttachmentPoint.GunRoot"/> is found;
-        /// otherwise, the <see cref="WeaponComponent"/> instance representing the invalid component.
-        /// </returns>
-        [Obsolete("WeaponComponentCollection.GetLuxuryFinishComponent is wrongly named and cannot necessarily get all of the components for gun_root (e.g. camo components)," +
-                  "use WeaponComponentCollection.GetGunRootComponent instead."),
-        EditorBrowsable(EditorBrowsableState.Never)]
-        public WeaponComponent GetLuxuryFinishComponent()
-        {
-            foreach (WeaponComponent component in this)
-            {
-                if (component.AttachmentPoint == WeaponAttachmentPoint.GunRoot)
                 {
                     return component;
                 }
